requirements.txt
# J.A.R.V.I.S. Python Dependencies
llama-cpp-python>=0.2.0
numpy>=1.24.0
matplotlib>=3.7.0
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99
protobuf>=4.22.0
fastapi>=0.100.0
uvicorn>=0.22.0
websockets>=11.0.0
python-multipart>=0.0.6
aiofiles>=23.0.0
pillow>=10.0.0
sounddevice>=0.4.6
pyaudio>=0.2.11
speechrecognition>=3.10.0
pyttsx3>=2.90
requests>=2.31.0
python-dotenv>=1.0.0
scipy>=1.10.0
imageio>=2.31.0
scikit-image>=0.21.0
imageio-ffmpeg>=0.4.8